move types into common file
